Moved ColorToNRGBA and unmultiplyAlpha into internal, fixed tests that created RGBAs
Moved ColorToNRGBA and unmultiplyAlpha into internal/color
Changed tests in markup_renderer_test that created RGBAs (which were failing now because of #2220 being fixed), so that they create NRGBAs instead.
