+++ dialog/color_picker.go
@@ -6,6 +6,7 @@ import (
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/canvas"
 	"fyne.io/fyne/v2/container"
+	col "fyne.io/fyne/v2/internal/color"
 	internalwidget "fyne.io/fyne/v2/internal/widget"
 	"fyne.io/fyne/v2/layout"
 	"fyne.io/fyne/v2/theme"
@@ -210,7 +211,7 @@ func (p *colorAdvancedPicker) CreateRenderer() fyne.WidgetRenderer {
 }
 
 func (p *colorAdvancedPicker) updateColor(color color.Color) bool {
-	r, g, b, a := theme.ColorToNRGBA(color)
+	r, g, b, a := col.ColorToNRGBA(color)
 	if p.Red == r && p.Green == g && p.Blue == b && p.Alpha == a {
 		return false
 	}
