+++ widget/entry_cursor_anim.go
@@ -7,6 +7,7 @@ import (
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/canvas"
+	col "fyne.io/fyne/v2/internal/color"
 	"fyne.io/fyne/v2/theme"
 )
 
@@ -29,7 +30,7 @@ func newEntryCursorAnimation(cursor *canvas.Rectangle) *entryCursorAnimation {
 // creates fyne animation
 func (a *entryCursorAnimation) createAnim(inverted bool) *fyne.Animation {
 	cursorOpaque := theme.PrimaryColor()
-	r, g, b, _ := theme.ColorToNRGBA(theme.PrimaryColor())
+	r, g, b, _ := col.ColorToNRGBA(theme.PrimaryColor())
 	cursorDim := color.NRGBA{R: uint8(r), G: uint8(g), B: uint8(b), A: 0x16}
 	start, end := color.Color(cursorDim), color.Color(cursorOpaque)
 	if inverted {
